Remove event changed from page to button component
