+++ charm-tracker/imports/ui/pages/index.js
@@ -5,7 +5,7 @@ export { default as EditVendors } from "./EditVendors"
 //export { default as Login } from "./Login"
 export { default as Logout } from "./Logout"
 export { default as PaymentManagement } from "./PaymentManagement"
-export { default as RemoveEvents } from "./RemoveEvents"
+// export { default as RemoveEvents } from "./RemoveEvents"
 export { default as Reports } from "./Reports"
 export { default as EditEvent } from "./EditEvent"
 //export { default as ResetPassword } from "./ResetPassword"
